requirements.txt
# Pinned ranges: upgrade with `pip install -U -r requirements.txt` then re-test.
requests>=2.33.0,<3
python-dotenv>=1.0.1
openai>=1.109.0,<2
supabase>=2.28.0,<3
psycopg2-binary>=2.9.9
tqdm>=4.66.0
pandas>=2.2.3,<3
numpy>=1.26.0,<3
rich>=13.7.0
tiktoken>=0.8.0
pyyaml>=6.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
PyJWT>=2.8.0
python-multipart>=0.0.26
# realtime (supabase) needs websockets.asyncio (13+) and <16 per realtime's constraint
websockets>=13,<16
